v1.1.1 Bugfix - production build crash
The electron-reload module normally instructs users to install it to your production dependencies.
Since it was not installed that way, but in devDependencies since I dont need it in production, the
import lines were not finding the module. Instead of moving to to production dependencies, I
just wrap the require in a try/catch statement like what is recommended in sindresorhus/electron-reloader