v1.1.2 Fix input price bug - replace innerhtml with value for input form
The app was getting into a bad state when a value in the inputPrice
field was updated. This is due to using innerHTML instead of value.
innerHtml is for editing the html attributes, which is sort of the same,
but not quite. Im not entirely sure why, but any actions afterwards would
result in updating a new set of files for example to break.

Price reverting also was fixed with this change.